+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+## 0.2.0
+- Added search filter
+- Added "reset" and "refresh" buttons
+- Added warnings when writing to mods.txt and mods.json
+- Added warning when trying to manage native UE4SS mods
+- Misc bug fixes
+
 ## 0.1.0
 - Initial version.